Moving EmitAsyncClosure to the separate file.
This beast is so huge, that I can't understand and fix it.
